Start refactor of the project structure
