+++ app/src/androidTest/java/com/github/fribourgsdp/radio/FireBaseTestActivityTest.kt
@@ -2,6 +2,7 @@ package com.github.fribourgsdp.radio
 
 import androidx.test.espresso.intent.Intents
 import androidx.test.ext.junit.rules.ActivityScenarioRule
+import com.github.fribourgsdp.radio.activities.testActivities.FireBaseTestActivity
 import org.junit.Rule
 import org.junit.Test
