+++ app/src/androidTest/java/com/github/fribourgsdp/radio/PlaylistTest.kt
@@ -1,5 +1,8 @@
 package com.github.fribourgsdp.radio
 
+import com.github.fribourgsdp.radio.backend.Genre
+import com.github.fribourgsdp.radio.backend.Playlist
+import com.github.fribourgsdp.radio.backend.Song
 import org.junit.Test
 
 import org.junit.Assert.*
