+++ app/src/androidTest/java/com/github/fribourgsdp/radio/SongTest.kt
@@ -1,5 +1,6 @@
 package com.github.fribourgsdp.radio
 
+import com.github.fribourgsdp.radio.backend.Song
 import org.junit.Test
 
 import org.junit.Assert.*
