+++ app/src/androidTest/java/com/github/fribourgsdp/radio/UserTest.kt
@@ -1,22 +1,14 @@
 package com.github.fribourgsdp.radio
 
 import android.content.Context
-import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
-import androidx.test.espresso.Espresso
-import androidx.test.espresso.action.ViewActions
-import androidx.test.espresso.intent.Intents
-import androidx.test.espresso.intent.matcher.IntentMatchers
-import androidx.test.espresso.matcher.ViewMatchers.withId
-import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
+import com.github.fribourgsdp.radio.backend.Genre
+import com.github.fribourgsdp.radio.backend.Playlist
+import com.github.fribourgsdp.radio.backend.User
 import junit.framework.Assert.assertEquals
-import org.hamcrest.Matchers
 import org.junit.Assert
-import org.junit.Before
 
-import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import java.lang.IllegalArgumentException
