+++ app/src/androidTest/java/com/github/fribourgsdp/radio/WorkingJoinGameActivityTest.kt
@@ -10,6 +10,7 @@ import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.fribourgsdp.radio.activities.*
 import com.github.fribourgsdp.radio.mockimplementations.*
 import org.hamcrest.Matchers
 import org.junit.After
