+++ app/src/androidTest/java/com/github/fribourgsdp/radio/WorkingLobbyActivityTest.kt
@@ -14,6 +14,7 @@ import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.fribourgsdp.radio.activities.*
 import org.hamcrest.Matchers.not
 import org.junit.Before
 import org.junit.After
