+++ app/src/main/java/com/github/fribourgsdp/radio/activities/GameSettingsActivity.kt
@@ -1,10 +1,14 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.activities
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import android.widget.*
+import com.github.fribourgsdp.radio.R
+import com.github.fribourgsdp.radio.activities.LobbyActivity
+import com.github.fribourgsdp.radio.backend.Playlist
+import com.github.fribourgsdp.radio.backend.User
 
 const val GAME_HOST_KEY = "com.github.fribourgsdp.radio.GAME_HOST"
 const val GAME_NAME_KEY = "com.github.fribourgsdp.radio.GAME_NAME"
