+++ app/src/main/java/com/github/fribourgsdp/radio/activities/SettingsActivity.kt
@@ -1,7 +1,8 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.activities
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.github.fribourgsdp.radio.R
 
 class SettingsActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
