+++ app/src/main/java/com/github/fribourgsdp/radio/activities/SpotifyReceiveActivity.kt
@@ -1,10 +1,12 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.activities
 
 import android.content.Intent
 import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
+//import com.github.fribourgsdp.radio.ImportSpotifyPlaylistsActivity
+import com.github.fribourgsdp.radio.R
 import com.spotify.sdk.android.auth.AuthorizationResponse
 
 class SpotifyReceiveActivity : AppCompatActivity() {
