+++ app/src/main/java/com/github/fribourgsdp/radio/activities/testActivities/DisplayLyricsActivity.kt
@@ -1,13 +1,12 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.activities.testActivities
 
-import android.content.Intent
 import android.os.Bundle
-import android.text.Html
-import android.view.View
 import android.widget.Button
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.text.HtmlCompat
+import com.github.fribourgsdp.radio.backend.LyricsGetter
+import com.github.fribourgsdp.radio.R
 import okhttp3.OkHttpClient
 
 
@@ -25,7 +24,11 @@ class DisplayLyricsActivity : AppCompatActivity() {
                 val artistName: String = artistTextView.text.toString()
                 val lyricsFuture = LyricsGetter.getLyrics(songName, artistName, OkHttpClient())
                 val lyrics = lyricsFuture.get()
-                resultsTextView.text = HtmlCompat.fromHtml(LyricsGetter.markSongName(lyrics, songName), HtmlCompat.FROM_HTML_MODE_LEGACY)
+                resultsTextView.text = HtmlCompat.fromHtml(
+                    LyricsGetter.markSongName(
+                        lyrics,
+                        songName
+                    ), HtmlCompat.FROM_HTML_MODE_LEGACY)
             }
         }
