+++ app/src/main/java/com/github/fribourgsdp/radio/backend/Database.kt
@@ -1,5 +1,6 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.backend
 
+import com.github.fribourgsdp.radio.backend.User
 import com.google.android.gms.tasks.Task
 import com.google.firebase.firestore.DocumentSnapshot
 import com.google.firebase.firestore.EventListener
