+++ app/src/main/java/com/github/fribourgsdp/radio/backend/StringComparisons.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.backend
 
 import java.text.Normalizer
