+++ app/src/main/java/com/github/fribourgsdp/radio/mockimplementations/LocalDatabase.kt
@@ -1,11 +1,11 @@
 package com.github.fribourgsdp.radio.mockimplementations
 
 import com.github.fribourgsdp.radio.*
+import com.github.fribourgsdp.radio.backend.*
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.firestore.DocumentSnapshot
 import com.google.firebase.firestore.EventListener
-import java.lang.IllegalArgumentException
 
 class LocalDatabase : Database {
     private val userMap: MutableMap<String, User> = mutableMapOf()
