+++ app/src/main/java/com/github/fribourgsdp/radio/mockimplementations/MockLobbyActivities.kt
@@ -1,7 +1,7 @@
 package com.github.fribourgsdp.radio.mockimplementations
 
-import com.github.fribourgsdp.radio.Database
-import com.github.fribourgsdp.radio.LobbyActivity
+import com.github.fribourgsdp.radio.backend.Database
+import com.github.fribourgsdp.radio.activities.LobbyActivity
 
 class WorkingLobbyActivity : LobbyActivity() {
     override fun initDatabase(): Database {
