+++ app/src/main/res/layout/activity_game_settings.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GameSettingsActivity">
+    tools:context="GameSettingsActivity">
 
     <TextView
         android:id="@+id/gameSettingsTitle"
