+++ app/src/main/res/layout/activity_google_sign_in.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="20dp"
-    tools:context=".GoogleSignInActivity">
+    tools:context="activities.GoogleSignInActivity">
 
 
     <TextView
