+++ app/src/main/res/layout/activity_lobby.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LobbyActivity">
+    tools:context=".activities.LobbyActivity">
 
     <TextView
         android:id="@+id/uuidText"
