+++ app/src/main/res/layout/activity_settings.xml
@@ -5,7 +5,7 @@
     android:id="@+id/volumeSlide"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SettingsActivity">
+    tools:context=".activities.SettingsActivity">
 
     <TextView
         android:id="@+id/settingsTitle"
